+++ resources/views/editor/blocks/footers/_columns.blade.php
@@ -4,10 +4,10 @@
 
     $redistribute = $redistribute ?? SiteFooterConfig::columnsRedistribute($config);
     $colStartClasses = [
-        1 => 'md:col-start-1',
-        2 => 'md:col-start-2',
-        3 => 'md:col-start-3',
-        4 => 'md:col-start-4',
+        1 => 'lg:col-start-1',
+        2 => 'lg:col-start-2',
+        3 => 'lg:col-start-3',
+        4 => 'lg:col-start-4',
     ];
 @endphp
 
@@ -21,7 +21,7 @@
     <div
         @class([
             'min-w-0 px-4 mb-10 w-full',
-            'basis-full md:mb-0 md:basis-0 md:flex-1' => $redistribute,
+            'basis-full lg:mb-0 lg:basis-0 lg:flex-1' => $redistribute,
             $colStartClasses[$index] => ! $redistribute && $columnVisible,
             $columnChrome['class'],
         ])