+++ resources/views/editor/blocks/footers/columns_brand_end.blade.php
@@ -1,9 +1,9 @@
 <div class="voodbuilder-editor-container px-5 py-24">
-    <div class="flex flex-col flex-wrap md:flex-row md:flex-nowrap md:items-center lg:items-start">
-        <div class="order-first flex grow flex-wrap md:order-none md:pr-12 md:text-left text-center">
+    <div class="flex flex-col flex-wrap lg:flex-row lg:flex-nowrap lg:items-start">
+        <div class="order-first flex grow flex-wrap lg:order-none lg:pr-12 lg:text-left text-center">
             @include('voodbuilder::editor.blocks.footers._columns', ['config' => $config])
         </div>
-        <div class="mx-auto mt-10 w-64 shrink-0 px-4 text-center md:mx-0 md:mt-0 md:text-left">
+        <div class="mx-auto mt-10 w-64 shrink-0 px-4 text-center lg:mx-0 lg:mt-0 lg:text-left">
             <div data-voodbuilder-brand></div>
             @include('voodbuilder::editor.blocks.partials.footer-tagline', [
                 'config' => $config,