+++ resources/views/editor/blocks/footers/columns_mission.blade.php
@@ -1,6 +1,6 @@
 <div class="voodbuilder-editor-container px-5 py-24">
-    <div class="flex flex-col flex-wrap md:flex-row md:flex-nowrap md:items-center lg:items-start">
-        <div class="mx-auto w-64 shrink-0 px-4 text-center md:mx-0 md:text-left">
+    <div class="flex flex-col flex-wrap lg:flex-row lg:flex-nowrap lg:items-start">
+        <div class="mx-auto w-64 shrink-0 px-4 text-center lg:mx-0 lg:text-left">
             <div data-voodbuilder-brand></div>
             @include('voodbuilder::editor.blocks.partials.footer-tagline', [
                 'config' => $config,
@@ -9,7 +9,7 @@
                 'taglineClass' => 'mt-3 text-sm text-vp-text-2',
             ])
         </div>
-        <div class="mt-10 flex grow flex-wrap md:mt-0 md:pl-12 md:text-left text-center">
+        <div class="mt-10 flex grow flex-wrap lg:mt-0 lg:pl-12 lg:text-left text-center">
             @include('voodbuilder::editor.blocks.footers._columns', ['config' => $config])
         </div>
     </div>