+++ resources/views/editor/blocks/footers/columns_newsletter.blade.php
@@ -4,7 +4,7 @@
     $newsletterChrome = SiteFooterConfig::chromeAttributes($config, 'newsletter', $preview);
     $redistribute = SiteFooterConfig::columnsRedistribute($config);
     $visibleMenuColumns = SiteFooterConfig::visibleFooterColumnIndexes($config);
-    $menuColumnsAlign = $redistribute && count($visibleMenuColumns) === 1 ? 'md:justify-end' : '';
+    $menuColumnsAlign = $redistribute && count($visibleMenuColumns) === 1 ? 'lg:justify-end' : '';
 @endphp
 
 <div class="voodbuilder-editor-container px-5 py-24">
@@ -38,12 +38,12 @@
             </button>
         </form>
     </div>
-    <div class="flex flex-col flex-wrap md:flex-row md:flex-nowrap md:items-start lg:items-start">
+    <div class="flex flex-col flex-wrap lg:flex-row lg:flex-nowrap lg:items-start">
         @include('voodbuilder::editor.blocks.footers._brand_column', ['config' => $config, 'preview' => $preview])
         <div @class([
-            'mt-10 w-full grow md:mt-0 md:pl-12 md:text-left text-center',
-            'flex flex-wrap md:flex-nowrap' => $redistribute,
-            'grid grid-cols-1 md:grid-cols-4' => ! $redistribute,
+            'mt-10 w-full grow lg:mt-0 lg:pl-12 lg:text-left text-center',
+            'flex flex-wrap lg:flex-nowrap' => $redistribute,
+            'grid grid-cols-1 lg:grid-cols-4' => ! $redistribute,
             $menuColumnsAlign,
         ]) data-voodbuilder-footer-menu-cols data-voodbuilder-footer-columns-redistribute="{{ $redistribute ? '1' : '0' }}">
             @include('voodbuilder::editor.blocks.footers._columns', [